feat: tanstack router setup done
