+++ src/router/index.tsx
@@ -1,0 +1,12 @@
+import { createRouter } from '@tanstack/react-router';
+import { routeTree } from './routes';
+
+const router = createRouter({ routeTree });
+
+declare module '@tanstack/react-router' {
+  interface Register {
+    router: typeof router;
+  }
+}
+
+export default router;
