+++ src/router/root.tsx
@@ -1,0 +1,8 @@
+import { createRootRoute } from '@tanstack/react-router';
+import BaseLayout from '../components/react-components/layouts/BaseLayout/BaseLayout';
+
+const rootRoute = createRootRoute({
+  component: BaseLayout,
+});
+
+export { rootRoute };
